requirements.txt
Django==3.1.3
djangorestframework
markdown
django-filter
pygments  
numpy
keras
nltk
django-environ
gunicorn
psycopg2-binary
whitenoise
django-cors-headers